requirements.txt
Django==4.1.5
djangorestframework==3.14.0
locustio==0.999
rest_framework_simplejwt==0.0.2
pillow==9.5.0
django-cors-headers==3.14.0
